+++ kittens/quick_access_terminal/main.go
@@ -74,6 +74,9 @@ func main(cmd *cli.Command, opts *Options, args []string) (rc int, err error) {
 	if conf.Start_as_hidden {
 		argv = append(argv, `--start-as-hidden`)
 	}
+	if conf.Grab_keyboard {
+		argv = append(argv, `--grab-keyboard`)
+	}
 	if conf.Hide_on_focus_loss {
 		argv = append(argv, `--hide-on-focus-loss`)
 	}